# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kritzeLLM API",
    "summary": "Search and chat with a handwritten journal.",
...
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Relative to wherever the app runs."
    }
  ],
  "paths": {
...
    "/journals": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "List journals",
        "description": "Lists all journals, newest first.",
        "operationId": "listJournals",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 25,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "description": "The `nextCursor` from the previous page. Omit for the first page.",
              "title": "Cursor"
            },
            "description": "The `nextCursor` from the previous page. Omit for the first page."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalList"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "journals"
        ],
        "summary": "Create a journal",
        "description": "Creates an empty journal to upload pages into.",
        "operationId": "createJournal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Journal"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URL of the new journal.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/journals/{journalId}": {
      "get": {
        "tags": [
          "journals"
        ],
        "summary": "Get a journal",
        "description": "Returns a single journal.",
        "operationId": "getJournal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "journalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^jrn_[0-7][0-9a-hjkmnp-tv-z]{25}$",
              "description": "The journal's ID.",
              "title": "Journalid"
            },
            "description": "The journal's ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Journal"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource doesn't exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "journals"
        ],
        "summary": "Update a journal",
        "description": "Renames a journal or changes its description. Only the given fields change.",
        "operationId": "updateJournal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "journalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^jrn_[0-7][0-9a-hjkmnp-tv-z]{25}$",
              "description": "The journal's ID.",
              "title": "Journalid"
            },
            "description": "The journal's ID."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Journal"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource doesn't exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "journals"
        ],
        "summary": "Delete a journal",
        "description": "Deletes a journal with all of its pages, images, blocks and search index entries.\n\nThis can't be undone.",
        "operationId": "deleteJournal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "journalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^jrn_[0-7][0-9a-hjkmnp-tv-z]{25}$",
              "description": "The journal's ID.",
              "title": "Journalid"
            },
            "description": "The journal's ID."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource doesn't exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/journals/{journalId}/pages": {
      "post": {
        "tags": [
          "journals",
          "pages"
        ],
        "summary": "Upload pages",
        "description": "Uploads page photos or scans into a journal.\n\nEach new page gets an `ingestPage` job that normalizes the image, transcribes it and indexes it.\n\nUploads are idempotent per file: a file whose SHA-256 already exists in this journal returns the\nexisting page with `duplicate: true` and no new job. Sending the same `Idempotency-Key` again\nreplays the original response.",
        "operationId": "uploadPages",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "journalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^jrn_[0-7][0-9a-hjkmnp-tv-z]{25}$",
              "description": "The journal's ID.",
              "title": "Journalid"
            },
            "description": "The journal's ID."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "description": "Any unique string (e.g. a UUID). Retries with the same key return the first response.",
              "title": "Idempotency-Key"
            },
            "description": "Any unique string (e.g. a UUID). Retries with the same key return the first response."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PageUpload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource doesn't exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "An uploaded file is too big.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "An uploaded file isn't a supported image type.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/journals/{journalId}/exports": {
      "post": {
        "tags": [
          "journals"
        ],
        "summary": "Export a journal",
        "description": "Starts a zipped export of all pages and transcriptions.\n\nWhen the job succeeds, download the file from its `resultUrl`.",
        "operationId": "exportJournal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "journalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^jrn_[0-7][0-9a-hjkmnp-tv-z]{25}$",
              "description": "The journal's ID.",
              "title": "Journalid"
            },
            "description": "The journal's ID."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URL of the job to poll.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "400": {
            "description": "The request couldn't be parsed (e.g. malformed JSON or multipart).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource doesn't exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Some values are invalid (including malformed IDs). Each one is listed in `errors`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportCreate": {
        "properties": {
          "format": {
            "$ref": "#/components/schemas/ExportFormat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "format"
        ],
        "title": "ExportCreate",
        "description": "Options for a journal export."
      },
      "ExportFormat": {
        "type": "string",
        "enum": [
          "markdown",
          "json"
        ],
        "title": "ExportFormat",
        "description": "`markdown`: one file per page plus images. `json`: everything machine-readable. Both come zipped."
      },
...
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^job_[0-7][0-9a-hjkmnp-tv-z]{25}$",
            "title": "Id",
            "description": "TypeID of a job."
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "target": {
            "$ref": "#/components/schemas/JobTarget"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobProgress"
              },
              {
                "type": "null"
              }
            ],
            "description": "How far along the job is, when it can tell."
          },
          "resultUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resulturl",
            "description": "Where to download the result once `succeeded` (exports only)."
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Problem"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why the job failed; `null` unless `status` is `failed`."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat",
            "description": "RFC 3339 timestamp in UTC."
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "description": "RFC 3339 timestamp in UTC."
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat",
            "description": "When a worker picked the job up."
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "description": "RFC 3339 timestamp in UTC."
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat",
            "description": "When the job succeeded or failed."
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "target",
          "progress",
          "resultUrl",
          "error",
          "createdAt",
          "startedAt",
          "finishedAt"
        ],
        "title": "Job",
        "description": "A unit of background work. Poll it until `status` is `succeeded` or `failed`."
      },
      "JobKind": {
        "type": "string",
        "enum": [
          "ingestPage",
          "transcribePage",
          "reindexPage",
          "reindexAll",
          "exportJournal"
        ],
        "title": "JobKind",
        "description": "What a job does."
      },
...
      "Journal": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^jrn_[0-7][0-9a-hjkmnp-tv-z]{25}$",
            "title": "Id",
            "description": "TypeID of a journal."
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "description": "Display name."
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional notes about the journal."
          },
          "pageCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pagecount",
            "description": "Number of pages in the journal."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat",
            "description": "RFC 3339 timestamp in UTC."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat",
            "description": "RFC 3339 timestamp in UTC."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "pageCount",
          "createdAt",
          "updatedAt"
        ],
        "title": "Journal",
        "description": "A notebook (or volume) of handwritten pages."
      },
      "JournalCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "description": "Display name."
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional notes about the journal."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "JournalCreate",
        "description": "A new journal."
      },
      "JournalList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Journal"
            },
            "type": "array",
            "title": "Items"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor",
            "description": "Pass as `cursor` to get the next page; `null` on the last page."
          }
        },
        "type": "object",
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "JournalList",
        "description": "A page of journals."
      },
      "JournalUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "New display name."
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "New notes; `null` clears them."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "JournalUpdate",
        "description": "Changes to a journal. Only the given fields change."
      },
...
      "PageUpload": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "Page images in page order. JPEG, PNG, HEIC/HEIF or WebP, up to 25 MB each."
          },
          "startPageNumber": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Startpagenumber",
            "description": "Page number for the first file; the rest count up from it. Defaults to the journal's next free page number."
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_uploadPages"
      },
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Identifies the problem type; `about:blank` when there's nothing more specific.",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "description": "The HTTP status code."
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "What went wrong in this particular case."
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "The request path that caused the problem."
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Individual validation problems (only for `422`)."
          }
        },
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "description": "An RFC 9457 problem details object."
      },
      "ProblemError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What's wrong with this value."
          },
          "pointer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pointer",
            "description": "JSON Pointer to the offending value in the request body, e.g. `/name`."
          },
          "parameter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameter",
            "description": "Name of the offending query, path or header parameter."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ProblemError",
        "description": "One invalid value in a request."
      },
...
      "UploadResult": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UploadItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "UploadResult",
        "description": "One entry per uploaded file, in upload order."
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "The `API_TOKEN` configured on the server. Only needed when one is set.",
        "scheme": "bearer"
      }
    }
  },
...
}